% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvFSGL.R
\name{cvFSGL}
\alias{cvFSGL}
\title{Fit and Cross-Validate a Linear Model with a Combination of Lasso and Group Lasso Regularization}
\usage{
cvFSGL(
  data,
  index,
  nfold = 10,
  nlam = 20,
  min.frac = 0.05,
  alpha = 0.95,
  lambdas = NULL,
  thresh = 1e-04,
  maxit = 10000,
  gamma = 0.8,
  verbose = TRUE,
  step = 1,
  reset = 10,
  foldid = NA,
  UseUpperBound = FALSE
)
}
\arguments{
\item{data}{data should be a list of $x$ and $y$, x is a data matrix (n x p) and y is a vector}

\item{index}{A p-vector indicating group membership of each covariate}

\item{nfold}{Number of folds of the cross-validation loop}

\item{nlam}{Number of lambda to use in the regularization path}

\item{min.frac}{The minimum value of the penalty parameter, as a fraction of the maximum value}

\item{alpha}{The mixing parameter. \code{alpha = 1} is the lasso penalty.}

\item{lambdas}{A user inputted sequence of lambda values for fitting. We recommend leaving this NULL and letting FSGL self-select values}

\item{thresh}{Convergence threshold for change in beta}

\item{maxit}{Maximum number of iterations to convergence}

\item{gamma}{Fitting parameter used for tuning backtracking (between 0 and 1)}

\item{verbose}{Logical flag for whether or not step number will be output}

\item{step}{Fitting parameter used for inital backtracking step size (between 0 and 1)}

\item{reset}{Fitting parameter used for taking advantage of local strong convexity in nesterov momentum (number of iterations before momentum term is reset)}

\item{foldid}{An optional user-pecified vector indicating the cross-validation fold in which each observation should be included. Values in this vector should range from 1 to nfold. If left unspecified, SGL will randomly assign observations to folds}

\item{UseUpperBound}{A logical flag for using upper bound}
}
\value{
An object with S3 class "cv.FSGL"
\describe{
\item{lldiff}{An \code{nlam} vector of cross validated negative log likelihoods (squared error loss in the linear case, along the regularization path)}
\item{llSD}{An \code{nlame} vector of approximate standard deviations of \code{lldiff}}
\item{lambdas}{The actual list of \code{lambda }values used in the regularization path.}
\item{fit}{A model fit object created by a call to \code{FSGL} on the entire dataset}
\item{foldid}{A vector indicating the cross-validation folds that each observation is assigned to}
\item{prevals}{A matrix of prevalidated predictions for each observation, for each lambda-value}
}
}
\description{
Fit and Cross-Validate a Linear Model with a Combination of Lasso and Group Lasso Regularization
}
\examples{
set.seed(1)
n = 50; p = 100; size.groups = 10
index <- ceiling(1:p / size.groups)
X = matrix(rnorm(n * p), ncol = p, nrow = n)
beta = (-2:2)
y = X[,1:5] \%*\% beta + 0.1*rnorm(n)
data = list(x = X, y = y)
cvFit = cvFSGL(data, index)
}
