% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSGL.R
\name{FSGL}
\alias{FSGL}
\title{FSGL}
\usage{
FSGL(
  data,
  index,
  maxit = 1000,
  thresh = 0.001,
  min.frac = 0.1,
  nlam = 20,
  gamma = 0.8,
  standardize = TRUE,
  verbose = FALSE,
  step = 1,
  reset = 10,
  alpha = 0.95,
  lambdas = NULL,
  UseUpperBound = FALSE
)
}
\arguments{
\item{data}{data should be a list of $x$ and $y$, x is a data matrix (n x p) and y is a vector}

\item{index}{A p-vector indicating group membership of each covariate}

\item{maxit}{Maximum number of iterations to convergence}

\item{thresh}{Convergence threshold for change in beta}

\item{min.frac}{The minimum value of the penalty parameter, as a fraction of the maximum value}

\item{nlam}{Number of lambda to use in the regularization path}

\item{gamma}{Fitting parameter used for tuning backtracking (between 0 and 1)}

\item{standardize}{Logical flag for variable standardization prior to fitting the model}

\item{verbose}{Logical flag for whether or not step number will be output}

\item{step}{Fitting parameter used for inital backtracking step size (between 0 and 1)}

\item{reset}{Fitting parameter used for taking advantage of local strong convexity in nesterov momentum (number of iterations before momentum term is reset)}

\item{alpha}{The mixing parameter. \code{alpha} = 1 is the lasso penalty. \code{alpha} = 0 is the group lasso penalty.}

\item{lambdas}{A user specified sequence of lambda values for fitting. We recommend leaving this NULL and letting SGL self-select values}

\item{UseUpperBound}{A logical flag for using error bound}
}
\value{
An object of type "FSGL"
\describe{
\item{beta}{A p by $nlam$ matrix, giving the penalized MLEs for the nlam different models, where the index corresponds to the penalty parameter $lambda$}
\item{lambdas}{The actual sequence of $lambda$ values used (penalty parameter)}
\item{X.transform}{A list used in $predict$ which gives the empirical mean and variance of the x matrix used to build the model}
}
}
\description{
Fit a regularized generalized linear model via penalized maximum likelihood. The model is fit for a path of values of the penalty parameter. Fits linear, logistic and Cox models.
}
\examples{
library(FSGL)

n = 100; p = 20; size.groups = 10
index <- ceiling(1:p / size.groups)
X = matrix(rnorm(n * p), ncol = p, nrow = n)
beta = (-2:2)
y = X[,1:5] \%*\% beta + 0.1*rnorm(n)
data = list(x = X, y = y)
fit = FSGL(data, index)
}
