% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKmatrix2d.R
\name{makeKmatrix2d}
\alias{makeKmatrix2d}
\title{Make 2D Fused K Matrix}
\usage{
makeKmatrix2d(dim1, dim2, groups)
}
\arguments{
\item{dim1}{number of rows of image.}

\item{dim2}{number of columns of image.}

\item{groups}{vector encoding group membership, with groups indexed as 1, 2, etc. Note: columns of image are concatenated from left to right, group membership coding should correspond.}
}
\description{
Builds the K matrix for fused sparse group lasso \code{fsgl.fit} function where graph structure is 2D fused. Requires Matrix package.
}
\examples{
makeKmatrix2d(dim1=3, dim2=3, groups=c(1, 2, 3, 2, 2, 1, 3, 3, 1))
}
\keyword{lasso}
