% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKmatrix3d.R
\name{makeKmatrix3d}
\alias{makeKmatrix3d}
\title{Make 3D Fused K Matrix}
\usage{
makeKmatrix3d(dim1, dim2, dim3, mask = NULL, groups)
}
\arguments{
\item{dim1}{dimension along x axis.}

\item{dim2}{dimension along y axis.}

\item{dim3}{dimension along z axis.}

\item{mask}{binary vector indicating masked areas (1s) zeros (0s).}

\item{groups}{vector encoding group membership, with groups indexed as 1, 2, etc. If there is a mask, put zeros or other placeholder at location of masked voxels. Note: columns of each slice of image are concatenated from left to right, group membership coding should correspond.}
}
\description{
Builds the K matrix for fused sparse group lasso \code{fsgl.fit} function where graph structure is 3D fused. Requires Matrix package.
}
\examples{
Kmatrix <- makeKmatrix3d(dim1=2, dim2=3, dim3=4, mask=c(rep(0, 4), rep(1, 16), rep(0, 4)), groups=c(rep(0, 4), rep(1, 4), rep(2, 4),  rep(3, 4), rep(4, 4), rep(0, 4)))
}
\keyword{lasso}
