% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Mask function to reduce dimension.}
\usage{
mask(data, mask)
}
\arguments{
\item{data}{a t-dimensional vector or a t by q dimensional matrix.}

\item{mask}{t-dimensional binary mask vector. Zeros indicate the entires or columns to be omitted, ones indicate entries or rows that will be kept.}
}
\description{
Transforms data from R^t to R^m, where t is the dimension of the total space and m is the dimension of the masked space. Input is the data (t-dimensional vector or a t by q dimensional matrix) to be transformed and a t-dimensional binary mask vector.
}
\examples{
x <- c(rep(0,4), 1:16, rep(0,4))
Mask <- c(rep(0, 4), rep(1, 16), rep(0, 4))
mask(x, Mask)
}
\keyword{mask}
