% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softthresh.R
\name{softthresh}
\alias{softthresh}
\title{Soft-thresholding function}
\usage{
softthresh(a, kappa)
}
\arguments{
\item{a}{a scalar or vector.}

\item{kappa}{soft-thresholding parameter.}
}
\description{
Soft-threshold a scalar or vector.
}
\examples{
softthresh(c(3, 4), 1)
}
\keyword{soft}
\keyword{thresholding}
