% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmask.R
\name{unmask}
\alias{unmask}
\title{Unmask function to return to original space.}
\usage{
unmask(data, mask)
}
\arguments{
\item{data}{a m-dimensional vector or a m by q dimensional matrix.}

\item{mask}{t-dimensional binary mask vector. Zeros indicate the entires or columns to be omitted, ones indicate entries or rows that will be kept.}
}
\description{
Transforms data from R^m to R^t, where m is the dimension of the masked space and t is the dimension of the total space. Input is the data (m-dimensional vector or a m by q dimensional matrix) to be transformed and a t-dimensional binary mask vector.
}
\examples{
x <- matrix(1:32, nrow=2)
Mask <- c(rep(0, 2), rep(1, 8), rep(0, 2), rep(1, 8), rep(0,4))
unmask(x, Mask)
}
\keyword{mask}
